add plot_gaussian_distributions function
The `plot_gaussian_distributions()` function is used to plot Gaussian distributions of the features in a dataset `X`, conditioned on the class labels `y`. The function visualizes the Gaussian distributions for each feature and class, along with a scatter plot of the data points colored by their corresponding class labels. It's a helper visualization function to illustrate the gaussian NB classifier